app/requirements.txt
torch==2.0.1
torchvision==0.15.2
transformers==4.30.2
pillow==9.5.0
numpy==1.24.3
pandas==1.5.3